Update API submodule to 1.2.0 upstream (#22)
* Update WiFi submodule to latest upstream (c3f1a3f)

Changes from 1.2.7 (c75a5e8) to c3f1a3f:
- begin() overloads now take const char* instead of char* (source-compatible)
- Adds getSocket(), firmwareVersion(), _state[], _server_port[] to WiFiClass
- WiFiClient: replaces single int-based constructor with WiFiClient() and
  WiFiClient(uint8_t sock); removes remoteIP/remotePort/localIP/localPort
- WiFiServer: removes destructor; replaces psock with pcb void*
- WiFiClass::isConnected() removed
- Adds utility/*.cpp/h SPI driver sources, wl_types.h, docs/, CI workflows

None of the removed APIs are used by ardulinux itself. Build and all 182
unit tests pass unchanged.

* Revert WiFi submodule to tagged 1.2.7 release

The tip of upstream master (c3f1a3f) contains only CI dependency bumps
via dependabot (GitHub Actions versions) — no library source changes.
Pinning to an untagged commit offers no benefit and makes the version
ambiguous. Stay on the 1.2.7 tag (c75a5e8), which is the latest stable
release.

* Update ArduinoCore-API to 1.2.0

API changes in 1.2.0:
- HardwareSPI: all virtual methods are now pure virtual (= 0) and a
  virtual destructor is added.
- HardwareI2C: requestFrom() return type changed from uint8_t to size_t.

Adaptations:
- LinuxHardwareI2C.h/cpp: requestFrom() return types updated to size_t.
- SimHardwareI2C.h: requestFrom() return types updated to size_t.
- HardwareSPIStubs.cpp: comment updated; the out-of-line definitions are
  still valid C++ for pure virtuals (callable via Base::method() syntax).
- test_spi_stubs.cpp: BareSPI now implements all pure virtuals by
  forwarding to the base-class stub bodies via HardwareSPI::method();
  the class is no longer abstract and the tests exercise the same paths.

Build clean, 182/182 unit tests pass.

+++ cores/ardulinux/linux/LinuxHardwareI2C.h
@@ -96,12 +96,12 @@ public:
    * @param len     Number of bytes to read.
    * @param stopBit Passed to the underlying I²C transaction (currently unused
    *                in the combined path).
-   * @return Number of bytes received, or a negative error code.
+   * @return Number of bytes received, or 0 on error.
    */
-  virtual uint8_t requestFrom(uint8_t address, size_t len, bool stopBit);
+  virtual size_t requestFrom(uint8_t address, size_t len, bool stopBit);
 
   /** Request @p len bytes from @p address with implicit STOP. */
-  virtual uint8_t requestFrom(uint8_t address, size_t len) {
+  virtual size_t requestFrom(uint8_t address, size_t len) {
     return requestFrom(address, len, true);
   }
